moved styling out of Navbar+AvatarDisplay
