+++ client/src/components/AvatarDisplay/AvatarDisplay.tsx
@@ -4,35 +4,12 @@ import { useState, MouseEvent } from 'react';
 import Menu from '@material-ui/core/Menu';
 import MenuItem from '@material-ui/core/MenuItem';
 import { useAuth } from '../../context/useAuthContext';
-import { makeStyles, createStyles, Theme } from '@material-ui/core/styles';
+import useStyles from './useStyles';
 
 interface Props {
   loggedIn: boolean;
   user: User;
 }
-
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    root: {
-      display: 'flex',
-      '& > *': {
-        margin: theme.spacing(1),
-      },
-    },
-    small: {
-      width: theme.spacing(3),
-      height: theme.spacing(3),
-    },
-    medium: {
-      width: theme.spacing(6),
-      height: theme.spacing(6),
-    },
-    large: {
-      width: theme.spacing(7),
-      height: theme.spacing(7),
-    },
-  }),
-);
 
 const AvatarDisplay = ({ user }: Props): JSX.Element => {
   const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
